Moved duplicate code to AbstractModel
